[Refactor] moved util code to separate header files
