+++ src/bliss/util/execute.h
@@ -1,0 +1,17 @@
+#ifndef EXECUTE_H
+#define EXECUTE_H
+#include <iostream>
+#include <random>
+#include <vector>
+
+#include "bliss/bliss_index.h"
+
+typedef unsigned long key_type;
+typedef unsigned long value_type;
+
+namespace bliss {
+namespace utils {
+namespace executor {}  // namespace executor
+}  // namespace utils
+}  // namespace bliss
+#endif
