+++ src/bliss_bench.cpp
@@ -10,85 +10,13 @@
 #include "bliss/bench_btree.h"
 #include "bliss/bench_lipp.h"
 #include "bliss/bliss_index.h"
+#include "bliss/util/args.h"
+#include "bliss/util/config.h"
+#include "bliss/util/execute.h"
 #include "bliss/util/reader.h"
 #include "bliss/util/timer.h"
 
-typedef unsigned long key_type;
-typedef unsigned long value_type;
-
-struct BlissConfig {
-    std::string data_file;
-    double preload_factor;
-    double write_factor;
-    double read_factor;
-    double mixed_read_write_ratio;
-    int seed;
-    int verbosity;
-    std::string index;
-    std::string file_type;
-    bool use_preload;
-};
-
-BlissConfig parse_args(int argc, char *argv[]) {
-    BlissConfig config;
-    cxxopts::Options options(
-        "bliss", "BLISS: Benchmarking Learned Index Structures for Sortedness");
-
-    try {
-        options.add_options()("d,data_file", "Path to the data file",
-                              cxxopts::value<std::string>())(
-            "p,preload_factor", "Preload factor",
-            cxxopts::value<double>()->default_value("0.5"))(
-            "w,write_factor", "Write factor",
-            cxxopts::value<double>()->default_value("0.25"))(
-            "r,read_factor", "Read factor",
-            cxxopts::value<double>()->default_value("0.1"))(
-            "m,mixed_read_write_ratio", "Read write ratio",
-            cxxopts::value<double>()->default_value("0.5"))(
-            "s,seed", "Random Seed value",
-            cxxopts::value<int>()->default_value("0"))(
-            "v,verbosity", "Verbosity [0: Info| 1: Debug | 2: Trace]",
-            cxxopts::value<int>()->default_value("0")->implicit_value("1"))(
-            "i,index", "Index type [alex | lipp | btree | bepstree | lsm]",
-            cxxopts::value<std::string>()->default_value("btree"))(
-            "file_type", "Input file type [binary | txt]",
-            cxxopts::value<std::string>()->default_value("txt"))(
-            "use_preload", "Use index defined preload",
-            cxxopts::value<bool>()->default_value("false"));
-
-        auto result = options.parse(argc, argv);
-        config = {
-            .data_file = result["data_file"].as<std::string>(),
-            .preload_factor = result["preload_factor"].as<double>(),
-            .write_factor = result["write_factor"].as<double>(),
-            .read_factor = result["read_factor"].as<double>(),
-            .mixed_read_write_ratio =
-                result["mixed_read_write_ratio"].as<double>(),
-            .seed = result["seed"].as<int>(),
-            .verbosity = result["verbosity"].as<int>(),
-            .index = result["index"].as<std::string>(),
-            .file_type = result["file_type"].as<std::string>(),
-            .use_preload = result["use_preload"].as<bool>(),
-        };
-    } catch (const std::exception &e) {
-        std::cerr << "Error: " << e.what() << std::endl;
-        std::cerr << options.help() << std::endl;
-        exit(1);
-    }
-    return config;
-}
-
-void display_config(BlissConfig config) {
-    spdlog::trace("Data File: {}", config.data_file);
-    spdlog::trace("Preload Factor: {}", config.preload_factor);
-    spdlog::trace("Write Factor: {}", config.write_factor);
-    spdlog::trace("Read Factor: {}", config.read_factor);
-    spdlog::trace("Read Write Ratio: {}", config.mixed_read_write_ratio);
-    spdlog::trace("Verbosity {}", config.verbosity);
-    spdlog::trace("Index: {}", config.index);
-    spdlog::trace("File type: {}", config.file_type);
-}
-
+using namespace bliss::utils;
 void execute_non_empty_reads(bliss::BlissIndex<key_type, value_type> &tree,
                              std::vector<key_type> &data, int num_reads,
                              int seed = 0) {
@@ -100,27 +28,6 @@ void execute_non_empty_reads(bliss::BlissIndex<key_type, value_type> &tree,
         key_idx = std::round(dist(gen) * (data.size() - 1));
         tree.get(data.at(key_idx));
     }
-}
-
-std::vector<std::pair<key_type, value_type>> create_preload_vec(
-    std::vector<key_type>::iterator &start,
-    std::vector<key_type>::iterator &end, bool sort_values = true,
-    int value_generator_seed = 0) {
-    std::mt19937 gen(value_generator_seed);
-    std::uniform_int_distribution<key_type> dist(0, 2 << 16);
-    std::vector<std::pair<key_type, value_type>> vec;
-
-    if (sort_values) {
-        spdlog::trace("Sorting values");
-        std::stable_sort(start, end);
-    }
-
-    spdlog::trace("Creating key-value pairs");
-    for (auto curr = start; curr != end; ++curr) {
-        vec.push_back(std::make_pair(*curr, dist(gen)));
-    }
-
-    return vec;
 }
 
 void execute_bulkload(bliss::BlissIndex<key_type, value_type> &tree,
@@ -170,10 +77,30 @@ void execute_mixed_workload(bliss::BlissIndex<key_type, key_type> &tree,
         }
     }
 }
+std::vector<std::pair<key_type, value_type>> create_preload_vec(
+    std::vector<key_type>::iterator &start,
+    std::vector<key_type>::iterator &end, bool sort_values = true,
+    int value_generator_seed = 0) {
+    std::mt19937 gen(value_generator_seed);
+    std::uniform_int_distribution<key_type> dist(0, 2 << 16);
+    std::vector<std::pair<key_type, value_type>> vec;
+
+    if (sort_values) {
+        spdlog::trace("Sorting values");
+        std::stable_sort(start, end);
+    }
+
+    spdlog::trace("Creating key-value pairs");
+    for (auto curr = start; curr != end; ++curr) {
+        vec.push_back(std::make_pair(*curr, dist(gen)));
+    }
+
+    return vec;
+}
 
 void workload_executor(bliss::BlissIndex<key_type, value_type> &tree,
-                       std::vector<key_type> &data, const BlissConfig &config,
-                       const int seed) {
+                       std::vector<key_type> &data,
+                       const config::BlissConfig &config, const int seed) {
     size_t num_inserts = data.size();
     size_t num_preload = std::round(config.preload_factor * num_inserts);
     size_t num_writes = std::round(config.write_factor * data.size());
@@ -233,7 +160,7 @@ void workload_executor(bliss::BlissIndex<key_type, value_type> &tree,
 }
 
 int main(int argc, char *argv[]) {
-    auto config = parse_args(argc, argv);
+    auto config = args::parse_args(argc, argv);
     switch (config.verbosity) {
         case 1:
             spdlog::set_level(spdlog::level::debug);
@@ -265,7 +192,7 @@ int main(int argc, char *argv[]) {
     } else if (config.index == "btree") {
         index.reset(new bliss::BlissBTreeIndex<key_type, value_type>());
     } else {
-        spdlog::error("{} not implemented yet", config.index);
+        spdlog::error(config.index + " not implemented yet", 1);
     }
 
     workload_executor(*index, data, config, 0);
